chore(trace-handler): test fixes to handle TrasePos changes [fixes VM-267]

+++ crates/air-lib/interpreter-data/src/executed_state/impls.rs
@@ -26,10 +26,10 @@ impl ParResult {
     }
 
     /// Returns a size of subtrace that this par describes in execution_step trace.
-    pub fn size(&self) -> Option<usize> {
+    pub fn size(&self) -> Option<u32> {
         self.left_size
             .checked_add(self.right_size)
-            .map(|v| v as usize)
+            .map(|v| v)
     }
 }
